feat: Add dock_atoms function for docking ship atoms to dock atoms
This commit adds a new function called `dock_atoms` to the `core.py` file. The `dock_atoms` function is used to dock ship atoms to dock atoms, with the default dock atoms being PF6. It calculates the vector between the centers of the ship atoms and dock atoms, applies an offset, and updates the positions of the dock atoms accordingly. Additionally, it includes an optional parameter for performing iMTD-GC sampling on the docked atoms. This function improves the functionality and flexibility of the code for docking atoms.